Move diary comments index action to comments controller
