+++ app/views/diary_comments/index.html.erb
@@ -27,8 +27,8 @@
   </table>
 
   <%= render "shared/pagination",
-             :newer_key => "diary_entries.comments.newer_comments",
-             :older_key => "diary_entries.comments.older_comments",
+             :newer_key => "diary_comments.index.newer_comments",
+             :older_key => "diary_comments.index.older_comments",
              :newer_id => @newer_comments_id,
              :older_id => @older_comments_id %>
 <% end -%>
