+++ config/locales/en.yml
@@ -592,7 +592,14 @@ en:
       all:
         title: "OpenStreetMap diary entries"
         description: "Recent diary entries from users of OpenStreetMap"
-    comments:
+    subscribe:
+      heading: Subscribe to the following diary entry discussion?
+      button: Subscribe to discussion
+    unsubscribe:
+      heading: Unsubscribe from the following diary entry discussion?
+      button: Unsubscribe from discussion
+  diary_comments:
+    index:
       title: "Diary Comments added by %{user}"
       heading: "%{user}'s Diary Comments"
       subheading_html: "Diary Comments added by %{user}"
@@ -602,12 +609,6 @@ en:
       comment: Comment
       newer_comments: "Newer Comments"
       older_comments: "Older Comments"
-    subscribe:
-      heading: Subscribe to the following diary entry discussion?
-      button: Subscribe to discussion
-    unsubscribe:
-      heading: Unsubscribe from the following diary entry discussion?
-      button: Unsubscribe from discussion
   doorkeeper:
     errors:
       messages:
